Turned initialAvailableOutgoingBitrate down to 10 Mbps 1 Gbps was responsible for very high probator data usage, and was unnecessary.
Renamed localConfig in server-core/src/config.ts to config, and removed existing config.
Existing config was not used, so having duplicates was just confusing - someone might
have changed the unused one and not seen their changes implemented. Had to cast import
as mediaConfig in some modules since appconfig.config was already imported as 'config'.

Fixed bug in selecting layer when switching to PiP; whether to use maxLayer-1 was based
on a hookstate object instead of its .value, so it always resolved truthy. But using max
layer is generally unnecessary given the size of UserMediaWindows, so made it use maxLayer-1
as long as the max height of user video is above a constant (currently set to 540p).