Updated media settings and fixed bugs. (#9709)
* Updated media settings and fixed bugs.

Added new client-settings for audio, video, and screenshare. For audio,
admins can set the max bitrate for opus (also hard-coded DTX to be on).
For video and screenshare, admins can select VP8, VP9, or h24 codec,
for video the maximum resolution that the user's webcam will get,
and for video and screenshare using simulcast (i.e. not VP9), the
maximum bitrates for the high/mid/low-end streams.

Fixed a bug where a lack of consumers would crash the instanceservers
since the network would be deleted. The correct check is if all
producers and consumers have been closed; seemed like a typo from
a recent conversion.

Fixed a bug in some useEffect listeners that were inadvertently closing
all consumers except the most recently created one.

* Made a single function for getting video/screenshare codec & encodings

* Renamed MediaProducerActions -> MediasoupMediaProducerActions

* Fixed bugs and linted